WGANGP

Documentation for WGANGP.

WGANGP.critic_loss

WGANGP.lipschitz1_gradient_loss

WGANGP.step_critic!

WGANGP.step_generator!

WGANGP.critic_loss — Method

critic_loss(m, x_true, x_generated, batch_size, λ)

WGAN-GP relaxed critic loss with lagrange multiplier λ

WGANGP.lipschitz1_gradient_loss — Method

lipschitz1_gradient_loss(m, x_true, x_generated, batch_size)

Estimates 𝐄ₓ(‖∇ₓD(x)‖₂ - 1)², where x is sampled uniformly on lines between points from the data distribution and the generators distribution

WGANGP.step_critic! — Method

step_critic!(opt, m, x_true, x_generated, batch_size; λ = 10.0f0)

A single optimisation step for the critic, with λ gradient penalty factor.

WGANGP.step_generator! — Method

A single optimisation step for the generator